libs/requirements.txt
watchdog
pyzmq
mongoengine
psutil